Avances

+++ Front-End/src/EditProfile.js
@@ -1,9 +1,11 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Button, Modal, ModalHeader, ModalBody, ModalFooter, Form, FormGroup, Label, Input } from 'reactstrap';
 import './EditProfile.css';
+import Login from './Login';
 
-function EditProfile() {
+function EditProfile() { 
+
   const [formData, setFormData] = useState({
     idParticipante: '',
     nombre: '',
@@ -13,14 +15,76 @@ function EditProfile() {
     gamertag: '',
     foto: null,
   });
-
+  
   const [modal, setModal] = useState(false);
   const [deleteModal, setDeleteModal] = useState(false);
   const [contrasenaEliminar, setContrasenaEliminar] = useState('');
+  const [deleteError, setDeleteError] = useState('');
+  const [errors, setErrors] = useState({});
+  const [isErrorModalOpen, setIsErrorModalOpen] = useState(false);
+  const [errorMessage, setErrorMessage] = useState('');
+  const [intentosFallidos, setIntentosFallidos] = useState(0);
+  const [successModalOpen, setSuccessModalOpen] = useState(false);
   const navigate = useNavigate();
 
   const toggleModal = () => setModal(!modal);
   const toggleDeleteModal = () => setDeleteModal(!deleteModal);
+
+  useEffect(() => {
+    // Obtener el ID del localStorage
+    const idParticipante = localStorage.getItem('id');
+    setFormData({ ...formData, idParticipante });
+  }, []);
+
+  if (!localStorage.getItem('tipo_usuario')) {
+    return <Login />
+  }
+
+  if (localStorage.getItem('tipo_usuario') !== 'participante') {
+    return('No tienes permisos para ver esta página.')
+  }
+
+  const datosValidos = () => {
+    let errors = {};
+    let isValid = true;
+
+    // Validación del correo
+    if(formData.correo){
+      if (typeof formData.correo !== "undefined") {
+        let lastAtPos = formData.correo.lastIndexOf('@');
+        let lastDotPos = formData.correo.lastIndexOf('.');
+
+        if (!(lastAtPos < lastDotPos && lastAtPos > 0 && formData.correo.indexOf('@@') === -1 && lastDotPos > 2 && (formData.correo.length - lastDotPos) > 2)) {
+          isValid = false;
+          errors["correo"] = "Correo inválido.";
+        }
+      }
+    }      
+    
+    // Validación de la contraseña
+    if(formData.contrasena){
+      if (formData.contrasena.length < 8) {
+        isValid = false;
+        errors["contrasena"] = "La contraseña debe tener al menos 8 caracteres.";
+      }
+    }
+
+     // Validación del gamertag
+    if (formData.gamertag) {
+      const regex = /^[a-zA-Z0-9]+$/; // Expresión regular para alfanuméricos sin caracteres especiales
+      const startsWithAt = formData.gamertag.startsWith('@');
+      const lengthValid = formData.gamertag.length >= 5;
+      const containsOnlyAlphanumeric = regex.test(formData.gamertag.slice(1));
+
+      if (!(startsWithAt && lengthValid && containsOnlyAlphanumeric)) {
+        isValid = false;
+        errors["gamertag"] = "Gamertag inválido. Debe comenzar con '@' y tener al menos 5 caracteres alfanuméricos.";
+      }          
+    } 
+
+    setErrors(errors);
+    return isValid;
+  };
 
   const handleFileChange = (e) => {
     const file = e.target.files[0];
@@ -32,7 +96,9 @@ function EditProfile() {
 
   const handleGuardarCambios = async (e) => {
     e.preventDefault();
-
+    if (!datosValidos()) {
+      return;
+    }
     console.log('Datos actualizados:', formData);
     
     try {
@@ -46,19 +112,70 @@ function EditProfile() {
 
       const data = await res.json();
       console.log('Respuesta del servidor:', data);
-      //toggleModal();
+      if (data.message === 'Perfil actualizado exitosamente') {       
+         // Verificar que los campos no estén vacíos
+        if (formData.nombre) {
+          localStorage.setItem('nombre', formData.nombre);
+        }
+        if (formData.apellido) {
+          localStorage.setItem('apellido', formData.apellido);
+        }
+
+        if (formData.correo) {
+          localStorage.setItem('email', formData.correo);
+        }
+
+        if (formData.contrasena) {
+          localStorage.setItem('contrasena', formData.contrasena);
+        }
+
+        if (formData.gamertag) {
+          localStorage.setItem('gamertag', formData.gamertag);
+        }
+
+        if (formData.foto) {
+          localStorage.setItem('foto', formData.foto);
+        }
+        openSuccessModal();  
+      } else if (data.error === 'Error, correo asociado a otra cuenta. Puede estar asociado a una cuenta no apta para participar.') {
+        setErrors({correo: data.error });
+      } else if (data.error === 'Error, tag ya asignado'){
+        setErrors({gamertag: data.error});
+      } else if (data.error) {
+        // Mostrar el error en el modal
+        setErrorMessage(data.error);
+        setIsErrorModalOpen(true);
+      }
     } catch (error) {
-      console.error('Error al enviar datos al servidor:', error);     
-    }
+      if (error.code === 'ECONNREFUSED') {
+        // Error de conexión rechazada (servidor no disponible)
+        console.error('Error: El servidor no está disponible');
+        // Mostrar el error en el modal
+        setErrorMessage('Error: El servidor no está disponible, intenta más tarde');
+        setIsErrorModalOpen(true);
+      } else {
+        // Otro tipo de error
+        console.error('Error al enviar datos al servidor:', error);
+        // Mostrar el error en el modal
+        setErrorMessage('Error en el servidor, intenta más tarde');
+        setIsErrorModalOpen(true);
+      }
+    }    
   };
 
   const handleEliminarPerfil = () => {
     toggleDeleteModal();
+    setDeleteError('');
   };
+  
 
   const handleConfirmarEliminar = async () => {
-    // Lógica para eliminar el perfil en el servidor
     try {
+      if (!contrasenaEliminar) {
+        setDeleteError('Ingresa tu contraseña');
+        return;
+      }
+      // Lógica para eliminar el perfil en el servidor
       const res = await fetch(`http://127.0.0.1:5000/participante/eliminarPerfil`, {
         method: 'DELETE',
         headers: {
@@ -71,34 +188,53 @@ function EditProfile() {
       });
   
       const data = await res.json();
-      console.log('Respuesta del servidor:', data); 
-      
+      console.log('Respuesta del servidor:', data);
   
-      // Cerrar el modal después de realizar la acción
-      toggleDeleteModal();
+      // Validar la respuesta del servidor
+      if (data.message === 'Perfil eliminado exitosamente') {
+        // Cerrar sesión y redirigir a la página de inicio
+        localStorage.clear();
+        navigate('/');
+      } else if (data.error === 'Contraseña incorrecta') {
+        // Aumentar el contador de intentos fallidos
+        setIntentosFallidos(intentosFallidos + 1);
+         // Mostrar el mensaje de error en el modal
+        if (intentosFallidos < 2) {
+          setDeleteError(`Contraseña incorrecta. Intentos restantes: ${2 - intentosFallidos}`);
+        } else {
+          setDeleteError('Demasiados intentos fallidos. Se cerrará la sesión.');
+        // Aquí puedes agregar la lógica para cerrar la sesión después de tres intentos fallidos
+          setTimeout(() => {
+          // Cerrar sesión y redirigir a la página de inicio
+            localStorage.clear();
+            navigate('/');
+          }, 3000);
+        }       
+      } else {      
+            // Actualizar el estado con el mensaje de error en caso de error de red
+        setDeleteError('Error de red, inténtalo de nuevo');
+      }
     } catch (error) {
-      console.error('Error al enviar datos al servidor:', error);      
+      console.error('Error al enviar datos al servidor:', error);
     }
   };
 
+  const handleCloseModal = () => {
+    setIsErrorModalOpen(false);
+  };
+
   const handleClickVolver = () => {
-    navigate(-1); // Navegar hacia atrás
+    navigate('/participante'); // Navegar hacia atrás
+  };
+
+  const openSuccessModal = () => {
+    setSuccessModalOpen(true);
   };
 
   return (
     <div className="EditProfile">
       <h1>Editar perfil</h1>
-      <form onSubmit={handleGuardarCambios}>
-
-      <FormGroup>
-          <Label for="idParticipante">ID participante:</Label>
-          <Input
-            type="number"
-            id="idParticipante"
-            value={formData.idParticipante}
-            onChange={(e) => setFormData({ ...formData, idParticipante: e.target.value })}          
-          />
-        </FormGroup>
+      <form onSubmit={handleGuardarCambios}>    
 
         <FormGroup>
           <Label for="nombre">Nombre:</Label>
@@ -108,6 +244,7 @@ function EditProfile() {
             value={formData.nombre}
             onChange={(e) => setFormData({ ...formData, nombre: e.target.value })}            
           />
+          {errors.nombre&& <div className="alert alert-danger">{errors.nombre}</div>}
         </FormGroup>
 
         <FormGroup>
@@ -118,6 +255,7 @@ function EditProfile() {
             value={formData.apellido}
             onChange={(e) => setFormData({ ...formData, apellido: e.target.value })}          
           />
+          {errors.apellido&& <div className="alert alert-danger">{errors.apellido}</div>}
         </FormGroup>
 
         <FormGroup>
@@ -128,6 +266,7 @@ function EditProfile() {
             value={formData.correo}
             onChange={(e) => setFormData({ ...formData, correo: e.target.value })}           
           />
+          {errors.correo&& <div className="alert alert-danger">{errors.correo}</div>}
         </FormGroup>
 
         <FormGroup>
@@ -138,6 +277,7 @@ function EditProfile() {
             value={formData.contrasena}
             onChange={(e) => setFormData({ ...formData, contrasena: e.target.value })}            
           />
+          {errors.contrasena&& <div className="alert alert-danger">{errors.contrasena}</div>}
         </FormGroup>
         
 
@@ -149,6 +289,7 @@ function EditProfile() {
             value={formData.gamertag}
             onChange={(e) => setFormData({ ...formData, gamertag: e.target.value })}            
           />
+          {errors.gamertag&& <div className="alert alert-danger">{errors.gamertag}</div>}
         </FormGroup>
 
         <FormGroup className="d-flex flex-column align-items-center">
@@ -164,7 +305,7 @@ function EditProfile() {
              Guardar cambios
             </Button>
             <Button color="primary" onClick={handleClickVolver}>
-             Volver
+             Volver a la vista de participante
             </Button>
             <Button color="danger" onClick={handleEliminarPerfil}>
             Eliminar perfil
@@ -215,6 +356,7 @@ function EditProfile() {
               />
             </FormGroup>
           </Form>
+          {deleteError && <div className="alert alert-danger">{deleteError}</div>}
         </ModalBody>
         <ModalFooter>
           <Button color="danger" onClick={handleConfirmarEliminar}>
@@ -225,6 +367,27 @@ function EditProfile() {
           </Button>
         </ModalFooter>
       </Modal>
+
+      <Modal isOpen={successModalOpen} toggle={() => setSuccessModalOpen(false)}>
+      <ModalHeader toggle={() => setSuccessModalOpen(false)}>Éxito</ModalHeader>
+        <ModalBody>
+          <p>Los datos han sido actualizados correctamente.</p>
+        </ModalBody>
+      <ModalFooter>
+          <Button color="primary" onClick={() => setSuccessModalOpen(false)}>
+            Cerrar
+          </Button>
+      </ModalFooter>
+      </Modal>
+
+      {isErrorModalOpen && (
+        <div className="modal">
+          <div className="modal-content">
+            <span className="close" onClick={handleCloseModal}>&times;</span>
+            <p>{errorMessage}</p>
+          </div>
+        </div>
+      )}
     </div>
   );
 }
